merges test helper modules
