+++ internal/codeintel/codenav/BUILD.bazel
@@ -76,7 +76,6 @@ go_test(
         "service_stencil_test.go",
         "service_syntactic_usages_test.go",
         "service_test.go",
-        "utils_test.go",
     ],
     embed = [":codenav"],
     tags = [TAG_PLATFORM_GRAPH],
